Backend/requirements.txt
databases~=0.4.3
PyJWT~=2.1.0
fastapi~=0.66.0
passlib~=1.7.4
pydantic~=1.8.2
uvicorn~=0.14.0
starlette~=0.14.2
asyncpg
python-multipart